Revert "optimise the common case by doing a pk_client_update_system if all the updates are selected"
This reverts commit 4ed2b769c5aa216db29ddcf635436ff63b0d191b.

mclasen: The way things are now, the "apply updates" button is backed by a
polkitgnomeaction that checks for the update-packages privilege, but you
now call a different action depending on the circumstances. The effect
is that first polkit-gnome asks me for the authorization for
update-packages, and then packagekit calls update-system, causing
another authentication dialog.

Duplicate authentication dialogs => bad