Convert all the client applications to use packagekit-glib2
This involves switching them to use GpkTask and a 100% async model.
There will be lots of bugs introduced with this commit, but hopefully
with glib2 the codebase will be more supportable and understandable.

Testing is very welcome, as we've only got 6 months to shake out all
the bugs from the conversion.