hide the update icon when all updates are blocked
Basically, gpk-update-icon tells the user about available updates even when
those updates are blocked. I investigated a bit and found that
PK_INFO_ENUM_BLOCKED isn't recognised as a package info field for GetUpdates.

I wrote a patch to resolve this, which works by setting PK_INFO_ENUM_BLOCKED to
the lowest-priority info field, and hiding the icon when it is returned from
pk_bitfield_contain_priority.

I hope you agree that this is a good solution. However, for some backends it
might be better to show a different icon rather than none at all, for example
when an update is blocked due to an error. I think, though, that temporarily
hiding an error (until a new update is found), is better than annoying a user
who may have explicitly blocked updates for some package.

Signed-off-by: Richard Hughes <richard@hughsie.com>