Ignore message types we do not recognize. Fixes rh#575669
If the daemon was built later than the client tools, the daemon
could send message types that the client does not recognize, and
does not have translations for. In this case ignore the message, as
we don't have anything useful (other than the enum number) to show
in the UI.

This should stop the message icon appearing and the when the user
clicks on it, they find the message is missing.

Thanks to Jcarr for debugging.