Fix an important bug that stopped GetUpdates from working correctly. Fixes rh#615099
We were sending an enum, when we should have been sending a bitfield.
All the time that PK_FILTER_ENUM_NONE was index zero this was no
problem, and it masked the bug. But when PK_FILTER_ENUM_UNKNOWN pushed
'none' into second spot in the enum list this caused a client side failure