Ensure we reset the GCancellable before doing each async method
If the user clicks the cancel button, then the cancellable gets set to cancelled state
and the transaction returns. If the user then clicks it again for another transaction
then nothing happens as we are already in the cancelled state.

libpackagekit-glib in git master is already patched to address this, and will return
an error if the method is called with a valid cancellable that is already in the
cancel state.