Pack package dependency lists into tabs
If all package dependency lists are showed in the dialog at the
same time, the window size of the dialog can easily become larger
than the screen, and the user can not click 'Cancel' or 'Continue'.
This patch packs the lists into several tabs, so the window size
is fixed and will not grow.

https://bugzilla.gnome.org/show_bug.cgi?id=653626

Signed-off-by: Gary Ching-Pang Lin <chingpang@gmail.com>
Signed-off-by: Richard Hughes <richard@hughsie.com>