trivial: move GSettings schemas to POTFILES.skip
Recent intltool update fixed a bug causing some GSettings schemas not
appearing in PO files. It broke statistics for string frozen release of
gnome-packagekit. As these strings are not visible to the user and it's
not like there's going to be any more 3.2.x releases, I've moved gschema
file from POTFILES.in to POTFILES.skip to ignore it in translation files.