Use a reasonable default window size
The insanely small previous default was probably from the days of
low-res displays. Since gpk-application currently always opens
with the default window size, you currently have to resize the
window every single time you open it, which sucks.

Nowadays gnome-shell handles low-res displays just fine. If you're
using an 800x600 resolution, this change causes the application to open
maximized (which is desirable for very small screens).

We additionally ought to remember our current size in gsettings when
we quit, but that does not excuse a poor default.

https://bugzilla.gnome.org/show_bug.cgi?id=701464