Set minimum content height for scrolledwindows
Previously, gtk+ would set a minimum height to display scroll bars. With
the overlay scrollbars in 3.16, gtk+ no longer does that and scrolled
windows end up with a minimum size of 1px by default.

The minimum size of 1px led the package details view to shrink to just a
single dotted line, with no way to resize it to make it readable; for
others such as the package view, they showed up in a right size, but it
was possible to resize the window and reduce them to just a dotted line.

Request a minimum content height of 100 pixels for all scrolledwindows
to work this around.