Update some group icons
This is still pretty bad, but at least some icons are shown correctly
again or have a more sensible symbol, as the icon had changed over time.

Ideally GPK would need to ship purpose-built icons, instead of using
some that were designed for other purposes, but this is good enough for
now.